requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1

# Data & Analysis
pandas==2.2.3
numpy==1.26.4

# Binance
python-binance==1.0.19

# Scheduler
apscheduler==3.10.4

# HTTP Client
httpx==0.27.2

# Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Logging
loguru==0.7.2

# Authentication
PyJWT==2.9.0
authlib==1.3.2
itsdangerous==2.2.0
passlib==1.7.4

# System Monitoring
psutil==6.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
